Move debounce to utilities
